Change type for size method to size_t

+++ yla_stack.c
@@ -99,7 +99,7 @@ int yla_stack_is_full(yla_stack* stack)
     return stack->count >= stack->size;
 }
 
-int yla_stack_count(yla_stack *stack)
+size_t yla_stack_count(yla_stack *stack)
 {
     return stack->count;
 }
